pyproject.toml
[tool.poetry]
name = "dapp"
version = "0.1.0"
description = ""
authors = ["fish895623 <dan990429@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.8"
django = "^4.1.7"
djangorestframework = "^3.14.0"


[tool.poetry.group.dev.dependencies]
black = "^23.1.0"
isort = "^5.12.0"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
